fixed benchmark, comparison tests
