--- test/simulate_100000/profile.json
@@ -1,183 +1,0 @@
-{
-	"ps": [
-		0.95,
-		0.9,
-		0.8,
-		0.7,
-		0.6,
-		0.5,
-		0.4,
-		0.3,
-		0.2,
-		0.1,
-		0.05
-	],
-	"normal": {
-		"speed": {
-			"km": [
-				58.02502679751552,
-				56.54679601687892,
-				54.762596965456666,
-				53.54728230670741,
-				52.49993925597806,
-				51.5188151471825,
-				50.55745751266199,
-				49.55303481211594,
-				48.397945259525564,
-				46.8581186822308,
-				45.549587615320576
-			],
-			"R4": [
-				54.48975957257346,
-				53.39113049800073,
-				52.01961274203623,
-				51.059346330275226,
-				50.2561662304703,
-				49.49295985620132,
-				48.72790956854004,
-				47.92054209598069,
-				46.989582025856656,
-				45.6054447137796,
-				44.48059259259259
-			],
-			"R5": [
-				54.99222184646601,
-				53.91130524913008,
-				52.58357247279225,
-				51.60031205304902,
-				50.76314471707561,
-				49.99274523423632,
-				49.22096164235548,
-				48.3883008356546,
-				47.416275346851656,
-				46.084846211552886,
-				44.97467620336362
-			],
-			"scipy": [
-				55,
-				54,
-				53,
-				52,
-				51,
-				50,
-				49,
-				48,
-				47,
-				46,
-				45
-			]
-		}
-	},
-	"poisson": {
-		"speed": {
-			"km": [
-				70.45315136504827,
-				66.63354968294117,
-				62.226878474083094,
-				59.12567854070487,
-				56.50944982050721,
-				54.203039625358954,
-				51.92805606412093,
-				49.55709621019116,
-				46.88412353736438,
-				43.314474109485126,
-				40.53037537626692
-			],
-			"R4": [
-				61.44289044289044,
-				58.68817651956703,
-				55.415823605706876,
-				53.127736450584486,
-				51.17728376611246,
-				49.382414882414885,
-				47.614508723599634,
-				45.743496672716276,
-				43.55376210578594,
-				40.629686896018555,
-				38.24304715840387
-			],
-			"R5": [
-				61.94955387783116,
-				59.20287253141831,
-				55.920348551791854,
-				53.63886597938144,
-				51.68259072117593,
-				49.882661996497376,
-				48.1116683084087,
-				46.23353640847113,
-				44.05068477115767,
-				41.11792618629174,
-				38.707903780068726
-			],
-			"scipy": [
-				62,
-				59,
-				56,
-				54,
-				52,
-				50,
-				48,
-				46,
-				44,
-				41,
-				39
-			]
-		}
-	},
-	"uniform": {
-		"speed": {
-			"km": [
-				69.68436516556295,
-				69.3687303311259,
-				68.63762713272534,
-				67.66375676915462,
-				66.27868490323742,
-				64.49266137719363,
-				62.32015799871493,
-				59.74151888322394,
-				56.71081066920009,
-				53.29599187290592,
-				51.44532317923004
-			],
-			"R4": [
-				68.50798141038594,
-				67.50090343304558,
-				65.51937524713325,
-				63.526032315978455,
-				61.53506854758594,
-				59.5307847082495,
-				57.53749747321609,
-				55.526305057096245,
-				53.518364105618424,
-				51.51411859620815,
-				50.51282051282051
-			],
-			"R5": [
-				69.01051370774555,
-				68.00090634441088,
-				66.01934465061193,
-				64.02607392607392,
-				62.03510342084328,
-				60.03076923076923,
-				58.037022253268134,
-				56.02600282201169,
-				54.018444666002,
-				52.01408734151741,
-				51.01291420562619
-			],
-			"scipy": [
-				69,
-				68,
-				66,
-				64,
-				62,
-				60,
-				58,
-				56,
-				54,
-				52,
-				51
-			]
-		}
-	}
-}
