+++ test/simulate_100000/summary.json
@@ -1,49 +1,43 @@
 {
-	"p": 0.85,
+	"p": 0.7,
 	"size": 100000,
 	"normal": {
 		"speed": {
 			"max": 64,
-			"km": 55.5145145070321,
-			"R4": 52.64422236102436,
-			"R5": 53.17094839712742,
-			"scipy": 53
+			"R4": 51.059346330275226,
+			"R5": 51.60031205304902,
+			"scipy": 52
 		},
 		"time": {
-			"km": 4.904506,
-			"R4": 0.120215,
-			"R5": 0.223739,
-			"scipy": 812.107388
+			"R4": 0.827916,
+			"R5": 0.177752,
+			"scipy": 798.61793
 		}
 	},
 	"poisson": {
 		"speed": {
 			"max": 82,
-			"km": 64.15190591052801,
-			"R4": 56.83908396946565,
-			"R5": 57.36439704675964,
-			"scipy": 57
+			"R4": 53.127736450584486,
+			"R5": 53.63886597938144,
+			"scipy": 54
 		},
 		"time": {
-			"km": 2.026859,
-			"R4": 0.057598,
-			"R5": 0.064651,
-			"scipy": 822.551407
+			"R4": 0.05854,
+			"R5": 0.055511,
+			"scipy": 821.339187
 		}
 	},
 	"uniform": {
 		"speed": {
 			"max": 70,
-			"km": 69.07271591178886,
-			"R4": 66.50630666141112,
-			"R5": 67.00636499254102,
-			"scipy": 67
+			"R4": 63.526032315978455,
+			"R5": 64.02607392607392,
+			"scipy": 64
 		},
 		"time": {
-			"km": 1.947613,
-			"R4": 0.029045,
-			"R5": 0.024719,
-			"scipy": 814.326249
+			"R4": 0.042681,
+			"R5": 0.030776,
+			"scipy": 777.818556
 		}
 	}
 }
