--- test/simulate_100000/summary2.json
@@ -1,54 +1,0 @@
-{
-	"size": 100000,
-	"normal": {
-		"speed": {
-			"max": 64,
-			"km_median": 51.52199063703076,
-			"km": 51.52309318039785,
-			"R4": 51.059346330275226,
-			"R5": 51.60031205304902,
-			"scipy": 52
-		},
-		"time": {
-			"km_median": 4.116381,
-			"km": 2.379821,
-			"R4": 0.211588,
-			"R5": 0.491367,
-			"scipy": 834.081112
-		}
-	},
-	"poisson": {
-		"speed": {
-			"max": 82,
-			"km_median": 54.20905581722251,
-			"km": 54.22296110922294,
-			"R4": 53.127736450584486,
-			"R5": 53.63886597938144,
-			"scipy": 54
-		},
-		"time": {
-			"km_median": 1.293009,
-			"km": 1.908231,
-			"R4": 0.098343,
-			"R5": 0.078459,
-			"scipy": 838.951268
-		}
-	},
-	"uniform": {
-		"speed": {
-			"max": 70,
-			"km_median": 64.49476614921949,
-			"km": 64.48223102998621,
-			"R4": 63.526032315978455,
-			"R5": 64.02607392607392,
-			"scipy": 64
-		},
-		"time": {
-			"km_median": 1.268785,
-			"km": 2.13473,
-			"R4": 0.042383,
-			"R5": 0.021244,
-			"scipy": 842.50671
-		}
-	}
-}
