--- test/simulate_20/profile.json
@@ -1,183 +1,0 @@
-{
-	"ps": [
-		0.95,
-		0.9,
-		0.8,
-		0.7,
-		0.6,
-		0.5,
-		0.4,
-		0.3,
-		0.2,
-		0.1,
-		0.05
-	],
-	"normal": {
-		"speed": {
-			"km": [
-				52.79668491725942,
-				52.59336983451884,
-				52.18673966903768,
-				51.43053528009963,
-				50.785931358665806,
-				50.42381775777038,
-				50.06170415687495,
-				48.4611632323237,
-				45.68550472975199,
-				43.31884217170239,
-				42.6139549813161
-			],
-			"R4": [
-				53,
-				52.5,
-				51,
-				50.6,
-				50.2,
-				49.333333333333336,
-				48,
-				47,
-				45,
-				43.5,
-				43
-			],
-			"R5": [
-				55,
-				53,
-				51.833333333333336,
-				51.166666666666664,
-				50.625,
-				50.125,
-				48.8,
-				47.333333333333336,
-				46,
-				44,
-				43.333333333333336
-			],
-			"scipy": [
-				54.5,
-				53,
-				51.5,
-				51,
-				51,
-				50,
-				49,
-				47.5,
-				46,
-				44,
-				43.5
-			]
-		}
-	},
-	"poisson": {
-		"speed": {
-			"km": [
-				59.815138932580126,
-				59.63027786516026,
-				59.26055573032051,
-				58.68082055345283,
-				57.63647546253682,
-				55.7613991222905,
-				52.1619988375954,
-				49.945521261101376,
-				48.67769137149245,
-				41.4917773911547,
-				38.672737210682904
-			],
-			"R4": [
-				60,
-				59,
-				57,
-				53,
-				52,
-				50.666666666666664,
-				49.5,
-				48.5,
-				46,
-				42,
-				40
-			],
-			"R5": [
-				61,
-				59.5,
-				57.5,
-				55,
-				52.6,
-				51.6,
-				50,
-				49,
-				47.333333333333336,
-				43.333333333333336,
-				41
-			],
-			"scipy": [
-				61,
-				59.5,
-				57.5,
-				54.5,
-				52.5,
-				52,
-				50,
-				49,
-				47,
-				44,
-				41
-			]
-		}
-	},
-	"uniform": {
-		"speed": {
-			"km": [
-				67.66561831530468,
-				67.33123663060938,
-				66.67889594085369,
-				66.04267187962792,
-				65.31727282931347,
-				62.93772796580725,
-				61.20406264671808,
-				58.17021414092498,
-				55.90128417019244,
-				55.085558227743576,
-				50.017999167549995
-			],
-			"R4": [
-				68,
-				67,
-				66,
-				65,
-				62,
-				60,
-				58,
-				56,
-				55.333333333333336,
-				51,
-				50
-			],
-			"R5": [
-				69,
-				67.66666666666667,
-				66.5,
-				65.33333333333333,
-				63,
-				60.5,
-				59,
-				56.75,
-				55.75,
-				53,
-				50.5
-			],
-			"scipy": [
-				69,
-				67.5,
-				66.5,
-				65.5,
-				63,
-				60.5,
-				59,
-				56.5,
-				56,
-				53,
-				50.5
-			]
-		}
-	}
-}
