+++ test/simulate_20/summary.json
@@ -1,49 +1,43 @@
 {
-	"p": 0.85,
+	"p": 0.7,
 	"size": 20,
 	"normal": {
 		"speed": {
 			"max": 56,
-			"km": 52.397003064194955,
-			"R4": 52,
-			"R5": 52.333333333333336,
-			"scipy": 52.5
+			"R4": 50.6,
+			"R5": 51.166666666666664,
+			"scipy": 51
 		},
 		"time": {
-			"km": 0.917378,
-			"R4": 0.101308,
-			"R5": 0.209587,
-			"scipy": 341.013531
+			"R4": 0.757295,
+			"R5": 0.15897,
+			"scipy": 319.976524
 		}
 	},
 	"poisson": {
 		"speed": {
 			"max": 62,
-			"km": 59.16359785449259,
-			"R4": 58,
-			"R5": 58.5,
-			"scipy": 58.5
+			"R4": 53,
+			"R5": 55,
+			"scipy": 54.5
 		},
 		"time": {
-			"km": 0.048809,
-			"R4": 0.021337,
-			"R5": 0.038738,
-			"scipy": 319.147231
+			"R4": 0.049919,
+			"R5": 0.079493,
+			"scipy": 352.348526
 		}
 	},
 	"uniform": {
 		"speed": {
 			"max": 70,
-			"km": 67.12833297928088,
-			"R4": 66.5,
-			"R5": 67,
-			"scipy": 67
+			"R4": 65,
+			"R5": 65.33333333333333,
+			"scipy": 65.5
 		},
 		"time": {
-			"km": 0.044094,
-			"R4": 0.020741,
-			"R5": 0.021652,
-			"scipy": 314.658872
+			"R4": 0.030533,
+			"R5": 0.028032,
+			"scipy": 320.463023
 		}
 	}
 }
